move common code to cli utils
