+++ ai-services/cmd/ai-services/cmd/catalog/uninstall.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/project-ai-services/ai-services/cmd/ai-services/cmd/catalog/common"
 	"github.com/project-ai-services/ai-services/internal/pkg/catalog/cli/uninstall"
-	catalogUtils "github.com/project-ai-services/ai-services/internal/pkg/catalog/utils"
+	cliutils "github.com/project-ai-services/ai-services/internal/pkg/catalog/cli/uninstall/utils"
 	"github.com/project-ai-services/ai-services/internal/pkg/vars"
 )
 
@@ -37,7 +37,7 @@ Examples:
 			return common.InitAndValidateRuntimeFlag(runtimeType)
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
-			return uninstall.Uninstall(catalogUtils.UninstallOptions{
+			return uninstall.Uninstall(cliutils.UninstallOptions{
 				Runtime:     vars.RuntimeFactory.GetRuntimeType(),
 				AutoYes:     uninstallAutoYes,
 				SkipCleanup: skipCleanup,
